+++ ShoppingApp/ShoppingApp/ViewModels/ProductCollectionViewCellViewModel.swift
@@ -27,15 +27,6 @@ final class ProductCollectionViewCellViewModel {
             completion(.failure(URLError(.badURL)))
             return
         }
-        
-        let request = URLRequest(url: url)
-        let task = URLSession.shared.dataTask(with: request) { data, _, error in
-            guard let data = data, error == nil else {
-                completion(.failure(error ?? URLError(.badServerResponse)))
-                return
-            }
-            completion(.success(data))
-        }
-        task.resume()
+        ImageLoader.shared.downLoadImage(url, completion: completion)
     }
 }
